fix(runtime): remove redundant LLM round cast

+++ crates/runtime/src/server/server_loop_host.rs
@@ -12950,7 +12950,7 @@ impl AgenticLoopHost for ServerAgenticLoopHost {
                 TurnPhaseOutcome::Succeeded,
             );
             if let Some(ref emitter) = state.messaging.progress_emitter {
-                emitter.llm_call_started(state.llm_rounds_completed as u32);
+                emitter.llm_call_started(state.llm_rounds_completed);
             } else {
                 self.emit_progress_event(json!({
                     "type": "agent_progress",
